+++ app/components/graph/chart-container-service.js
@@ -5,6 +5,12 @@
 angular.module('lizard-nxt')
   .factory('ChartContainer', ['NxtD3', function (NxtD3) {
 
+  var DEFAULT_GREEN = '#16a085';
+
+  var defaultKeys = {
+    x: 'timestamp',
+    y: { 'y0': 'min', 'y1': 'max' }
+  };
 
   /**
    * Charts are are small objects that are drawn on the graph canvas
@@ -19,55 +25,22 @@ angular.module('lizard-nxt')
    *
    * For now this is a way to keep track of the scales, domains and xy's of the graph
    */
-  function ChartContainer (content, graph, temporal) {
+  function ChartContainer (chartContent) {
 
-    var DEFAULT_GREEN = '#16a085';
+    this.id = chartContent.id;
+    this.keys = chartContent.keys || defaultKeys;
+    this.color = chartContent.color || DEFAULT_GREEN;
+    this.unit = chartContent.unit;
+    this.setContentUpdateY(chartContent);
 
-    this._x = null;
-
-    this._graph = graph;
-
-    this.data = content.data;
-    var defaultKeys = {
-      x: 'timestamp',
-      y: { 'y0': 'min', 'y1': 'max' }
-    };
-    this.keys = content.keys || defaultKeys;
-    this.labels = content.labels || {x: '', y:''};
-
-    this.color = content.color || DEFAULT_GREEN;
-
-    var options = {
-      x: {
-        scale: 'time',
-        orientation: 'bottom'
-      },
-      y: {
-        scale: 'linear',
-        orientation: 'left'
-      }
-    };
-    this._xy = this._graph._createXYGraph(
-      this.data,
-      this.keys,
-      this.labels,
-      temporal ? options : undefined,
-      this._graph._xDomainInfo
-    );
     return;
   }
 
-  ChartContainer.prototype.updateXY = function (content) {
-    this.data = content.data;
-    this._xy = this._graph.rescale(
-      this._graph._svg,
-      this._graph.dimensions,
-      this._xy,
-      this.data,
-      this.keys,
-      null,
-      this._graph._xDomainInfo
-    );
+  ChartContainer.prototype.setContentUpdateY = function (chartContent) {
+    this.data = chartContent.data;
+    this.keys = chartContent.keys || defaultKeys;
+    this.color = chartContent.color || DEFAULT_GREEN;
+    this.yMaxMin = NxtD3.prototype._maxMin(this.data, this.keys.y);
   };
 
   return ChartContainer;